pyproject.toml
[project]
name = "airflow-mcp-iap"
version = "0.1.0"
description = "Apache Airflow MCP Server with Google IAP authentication"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "apache-airflow-client>=2.10.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.2.4",
    "requests>=2.32.5",
]

[project.scripts]
airflow-mcp-iap = "airflow_mcp_iap.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
